pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "torchkrates"
version = "0.1"
description = "Implementation of So3krates in pytorch"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Placeholder" }
authors = [
    { name = "Tobias Henkes", email = "tobias.henkes@uni.lu" }
]
keywords = []
classifiers = [
    "Programming Language :: Python :: 3"
]
dependencies = [
    "ase",
    "numpy",
    "mace-torch",
    "torch",
    "pyYaml",
]

[project.urls]
homepage = "https://github.com/tohenkes/So3krates-torch"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
so3krates_torch = ["pretrained/so3lr/so3lr.model"]

[project.scripts]
torchkrates-jax2torch = "so3krates_torch.cli.jax_to_torch:main"
torchkrates-torch2jax = "so3krates_torch.cli.torch_to_jax:main"
torchkrates-eval = "so3krates_torch.cli.run_eval:main"
torchkrates-test = "so3krates_torch.cli.run_test:main"